split server.cpp
