+++ Makefile
@@ -8,7 +8,8 @@ BUILD_DIR := build
 SRC_DIR := src
 
 MAIN		:= main.cpp
-SERVER		:= Server.cpp Client.cpp Response.cpp ResponseUtils.cpp ResponseErrors.cpp	\
+SERVER		:= Server.cpp Server_connections.cpp Server_io.cpp Server_dispatch.cpp \
+				Client.cpp Response.cpp ResponseUtils.cpp ResponseErrors.cpp	\
 				MultipartParser.cpp ResponseHandlers.cpp DefaultError.cpp Session.cpp	\
 				AutoindexResponse.cpp
 CONFIG		:= WebserverConfig.cpp ServerConfig.cpp RouteRule_CGI.cpp PathPattern.cpp ConfigError.cpp
