+++ src/server/Server.cpp
@@ -1,405 +1,9 @@
 #include "Server.hpp"
 #include "Response.hpp"
-#include "ResponseHandlers.hpp"
 #include "DefaultError.hpp"
 #include <cstddef>
 #include <ctime>
 #include <vector>
-
-void Server::dispatch_request(const FileDescriptor *client_fd,
-                              ClientSession &client, char **envp) {
-  RouteRule_CGI const *cgi_path = client.config->find_route_cgi(
-      client.req->get_method(), client.req->get_path());
-
-  // Log request details
-  Result<size_t> content_length = client.req->get_content_length();
-  std::cout << "\n" << utils::info << "client ip: " << client.ip << std::endl;
-  std::cout << utils::info << "[Request] " << client.req->get_method_string()
-            << " " << client.req->get_path() << " (Body: ";
-  if (content_length.has_value())
-    std::cout << content_length.value();
-  else
-    std::cout << "(non-existent)";
-  std::cout << " bytes)" << std::endl;
-  for (std::map<std::string, std::string>::const_iterator it =
-           client.req->get_headers().begin();
-       it != client.req->get_headers().end(); ++it) {
-    std::cout << utils::info << "[Request]" << it->first << ": " << it->second
-              << std::endl;
-  }
-
-  // Dispatch to CGI if applicable
-  if (cgi_path != NULL) {
-    Result<Void> del_ = ResponseHandlers::register_cgi(
-        *client.req, *cgi_path, &epoll, config.get_global_cgi(), cgis,
-        client_fd, envp);
-    if (!del_.has_value())
-      std::cerr << utils::error << "CGI registration failed: " << del_.error()
-                << std::endl;
-    delete client.req;
-    client.req = NULL;
-    return;
-  }
-
-  // Generate normal HTTP response
-  Response http = ResponseHandlers::http_response(client.req, &client, mime_type,
-                                                   &sessions, envp);
-  http.print_simple(std::cout);
-
-  delete client.req;
-  client.req = NULL;
-
-  queue_response(client_fd, http);
-}
-
-void Server::queue_response(const FileDescriptor *client_fd,
-                            const Response &response) {
-  if (clients.find(client_fd) == clients.end())
-    return;
-  ClientSession &client = clients.at(client_fd);
-
-  const size_t STREAM_THRESHOLD = 64 * 1024; // 64 KiB
-
-  if (!response.file_path.empty()) {
-    // For small files, buffer entirely; for large files, stream from disk.
-    if (response.content_length <= STREAM_THRESHOLD) {
-      std::ifstream infile(response.file_path.c_str(), std::ios::binary);
-      if (infile.is_open()) {
-        std::ostringstream ss;
-        ss << response; // headers
-        ss << infile.rdbuf();
-        client.out_buff += ss.str();
-        infile.close();
-      } else {
-        // Fallback: send 404-like error
-        Response err = DefaultError::default_err_response(Response::NOT_FOUND);
-        std::ostringstream ss;
-        ss << err;
-        client.out_buff += ss.str();
-        client.dropping = true;
-      }
-    } else {
-      // Stream: send headers now, keep file open to stream body later
-      std::ostringstream ss;
-      ss << response; // headers only (body is not set)
-      client.out_buff += ss.str();
-      client.out_file_path = response.file_path;
-      client.out_file_offset = 0;
-      if (!client.out_file_path.empty()) {
-        client.streaming_file = true;
-      } else {
-        // Could not open: send 404
-        Response err = DefaultError::default_err_response(Response::NOT_FOUND);
-        std::ostringstream es;
-        es << err;
-        client.out_buff += es.str();
-        client.dropping = true;
-      }
-    }
-  } else {
-    std::ostringstream oss;
-    oss << response;
-    client.out_buff += oss.str();
-  }
-
-  // Mark for disconnection if client doesn't want keep-alive
-  if (!response.keep_alive)
-    client.dropping = true;
-}
-
-void Server::new_connection(const FileDescriptor *server_fd) {
-  while (true) { // accept all clients until nothing to connect
-    // init client socket
-    sockaddr_in client_addr = {};
-    socklen_t client_len = sizeof(client_addr);
-    Result<FileDescriptor> client_result = server_fd->socket_accept(
-        reinterpret_cast<struct sockaddr *>(&client_addr), &client_len);
-    if (!client_result.has_value()) {
-      const std::string &err = client_result.error();
-      if (err == Errors::try_again)
-        break; // EWOULDBLOCK: nothing to connect
-      else if (err == Errors::interrupted)
-        continue; // EINTR: accept retry
-      else {
-        std::cerr << utils::error << "accept failed: " << err << std::endl;
-        break;
-      }
-    }
-    FileDescriptor client_fd = client_result.value();
-    if (!client_fd.set_nonblocking().has_value()) {
-      std::cerr << utils::error
-                << "failed to set client socket to non-blocking mode"
-                << std::endl;
-      continue;
-    }
-
-    if (!client_fd.close_on_exec().has_value()) {
-      std::cerr << utils::error
-                << "failed to set client socket to close-on-exec mode"
-                << std::endl;
-      continue;
-    }
-
-    ClientSession client;
-    if (clock_gettime(CLOCK_MONOTONIC, &client.last_activity_time) != 0) {
-      std::cerr << utils::error
-                << "failed to get current time for client activity"
-                << std::endl;
-      continue;
-    }
-    std::ostringstream oss;
-    const unsigned char *octets =
-        reinterpret_cast<unsigned char *>(&client_addr.sin_addr.s_addr);
-    oss << static_cast<unsigned int>(octets[0]) << '.'
-        << static_cast<unsigned int>(octets[1]) << '.'
-        << static_cast<unsigned int>(octets[2]) << '.'
-        << static_cast<unsigned int>(octets[3]);
-    client.ip = oss.str();
-
-    // register client socket to EPoll
-    Event client_event(NULL, true, true, true, false, false, false);
-    Option client_option(true, false, false, false);
-
-    Result<FileDescriptor *> add_result =
-        epoll.add_fd(client_fd, client_event, client_option);
-    if (add_result.has_value()) {
-      const FileDescriptor *client_ptr = add_result.value();
-      if (listeners.find(server_fd) != listeners.end()) {
-        client.config = listeners.at(server_fd);
-      }
-      clients[client_ptr] = client;
-      std::cout << utils::info << "New client connected!" << std::endl;
-    } else {
-      std::cerr << utils::error << "epoll add failed: " << add_result.error()
-                << std::endl;
-    }
-  }
-}
-
-void Server::disconnect(const FileDescriptor *client_fd) {
-  epoll.del_fd(client_fd);
-  clients.erase(client_fd);
-}
-
-void Server::client_read(const FileDescriptor *client_fd, char **envp) {
-  if (clients.find(client_fd) == clients.end()) {
-    return;
-  }
-  ClientSession &client = clients.at(client_fd);
-  if (client.config == NULL) {
-    Response resp(
-        DefaultError::default_err_response(Response::INTERNAL_SERVER_ERR));
-    resp.print_simple(std::cout);
-    queue_response(client_fd, resp);
-    return;
-  }
-  if (clock_gettime(CLOCK_MONOTONIC, &client.last_activity_time) != 0) {
-    std::cerr << utils::error << "failed to update client activity time"
-              << std::endl;
-    return;
-  }
-  while (true) { // repeat until nothing to read
-    char buf[NETWORK_BUFFER_SIZE];
-    Result<ssize_t> recv_res = client_fd->sock_recv(buf, sizeof(buf));
-    if (!recv_res.has_value())
-      break; // EWOULDBLOCK
-
-    if (clock_gettime(CLOCK_MONOTONIC, &client.last_activity_time) != 0) {
-      std::cerr << utils::error << "failed to update client activity time"
-                << std::endl;
-      return;
-    }
-    ssize_t bytes = recv_res.value();
-    if (bytes == 0) { // (EOF)
-      client.dropping = true;
-      break;
-    }
-    client.in_buff.append(buf, static_cast<std::size_t>(bytes));
-  }
-
-  // HTTP parsing and response generate
-  std::string &in_buffer = client.in_buff;
-  while (!in_buffer.empty()) {
-    if (client.req == NULL) {
-      Result<Request *> req_ = Request::from_buff(in_buffer);
-
-      if (!req_.has_value()) {
-        std::cerr << utils::error << "request parsing failed: " << req_.error()
-                  << std::endl;
-        if (req_.error() == Errors::incomplete_header) {
-          if (client.dropping)
-            disconnect(client_fd);
-          return;
-        } else if (req_.error() == Errors::malformed_header ||
-                   req_.error() == Errors::bad_request) {
-          Response resp(
-              DefaultError::default_err_response(Response::BAD_REQUEST));
-          resp.headers = client.config->get_header();
-          resp.print_simple(std::cout);
-          queue_response(client_fd, resp);
-          return;
-        } else if (req_.error() == Errors::not_implemented) {
-          Response resp =
-              DefaultError::default_err_response(Response::NOT_IMPLEMENTED);
-          resp.headers = client.config->get_header();
-          resp.print_simple(std::cout);
-          queue_response(client_fd, resp);
-          return;
-        }
-      }
-
-      client.req = req_.value();
-      const Result<size_t> req_cl = client.req->get_content_length();
-      const RouteRule *rule = client.config->find_route(
-          client.req->get_method(), client.req->get_path());
-      if (rule != NULL && req_cl.has_value() &&
-          req_cl.value() > static_cast<size_t>(rule->max_body_KB) * 1024) {
-        Response resp(
-            DefaultError::default_err_response(Response::PAYLOAD_TOO_LARGE));
-        resp.headers = client.config->get_header();
-        resp.print_simple(std::cout);
-        queue_response(client_fd, resp);
-        return;
-      }
-      if (client.req->is_partial()) {
-        if (client.dropping)
-          disconnect(client_fd);
-        return;
-      }
-
-      dispatch_request(client_fd, client, envp);
-      // Client may have been disconnected in dispatch_request, check existence
-      if (clients.find(client_fd) == clients.end())
-        return;
-      continue;
-    } else {
-      // Refresh client reference before continuing
-      if (clients.find(client_fd) == clients.end())
-        return;
-      const Result<Void> parse_more = client.req->continue_parsing(in_buffer);
-      if (!parse_more.has_value()) {
-        std::cerr << utils::error
-                  << "request body parsing failed: " << parse_more.error()
-                  << std::endl;
-        Response resp(
-            DefaultError::default_err_response(Response::BAD_REQUEST));
-        resp.headers = client.config->get_header();
-        resp.print_simple(std::cout);
-        queue_response(client_fd, resp);
-        delete client.req;
-        client.req = NULL;
-        return;
-      }
-      const Result<size_t> content_len = client.req->get_content_length();
-      const RouteRule *const rule = client.config->find_route(
-          client.req->get_method(), client.req->get_path());
-      if (content_len.has_value() && rule != NULL &&
-          (static_cast<size_t>(rule->max_body_KB) * 1024 <
-               content_len.value() ||
-           (client.req->is_partial() &&
-            content_len.value() <= client.req->get_body().size()) ||
-           (!client.req->is_partial() &&
-            content_len.value() < client.req->get_body().size()))) {
-        Response resp(
-            DefaultError::default_err_response(Response::PAYLOAD_TOO_LARGE));
-        resp.headers = client.config->get_header();
-        resp.print_simple(std::cout);
-        queue_response(client_fd, resp);
-        delete client.req;
-        client.req = NULL;
-        return;
-      }
-      if (client.req->is_partial())
-        return;
-
-      dispatch_request(client_fd, client, envp);
-      // Client may have been disconnected in dispatch_request, check existence
-      if (clients.find(client_fd) == clients.end())
-        return;
-      continue;
-    }
-  }
-  client_write(client_fd); // when the response is generated freshly, likely
-                           // EPOLLIN | EPOLLOUT
-}
-
-void Server::client_write(const FileDescriptor *client_fd) {
-  if (clients.find(client_fd) == clients.end())
-    return;
-  ClientSession &client = clients.at(client_fd);
-  if (clock_gettime(CLOCK_MONOTONIC, &client.last_activity_time) != 0) {
-    std::cerr << utils::error << "failed to update client activity time"
-              << std::endl;
-    return;
-  }
-  std::string &write_buffer = client.out_buff;
-  if (!write_buffer.empty()) {
-    while (true) { // ET 모드이므로 보낼 수 있는 만큼 다 보냄
-      Result<ssize_t> send_res =
-          client_fd->sock_send(write_buffer.c_str(), write_buffer.length());
-      if (!send_res.has_value())
-        break; // EWOULDBLOCK
-
-      const ssize_t bytes = send_res.value();
-      if (bytes == 0)
-        break;
-
-      write_buffer.erase(0, static_cast<std::size_t>(bytes));
-      if (write_buffer.empty()) {
-        if (client.dropping)
-          disconnect(client_fd);
-        return;
-      }
-    }
-  } else if (client.dropping) {
-    disconnect(client_fd);
-    return;
-  }
-
-  // If we've emptied the write buffer and are streaming a file, load next
-  // chunk from disk into the buffer and attempt to send it immediately.
-  if (write_buffer.empty() && client.streaming_file) {
-    const size_t CHUNK = NETWORK_BUFFER_SIZE; // 4096
-    std::vector<char> buf(CHUNK);
-    if (client.out_file_path.empty()) {
-      client.streaming_file = false;
-    } else {
-      std::ifstream infile(client.out_file_path.c_str(), std::ios::binary);
-      if (!infile.is_open()) {
-        client.streaming_file = false;
-        if (client.dropping)
-          disconnect(client_fd);
-        return;
-      }
-      infile.seekg(static_cast<std::streamoff>(client.out_file_offset));
-      infile.read(buf.data(), static_cast<std::streamsize>(CHUNK));
-      std::streamsize read_bytes = infile.gcount();
-      if (read_bytes > 0) {
-        client.out_file_offset += static_cast<size_t>(read_bytes);
-        client.out_buff.append(buf.data(), static_cast<size_t>(read_bytes));
-        // Try sending what we just read
-        while (!client.out_buff.empty()) {
-          Result<ssize_t> send_res = client_fd->sock_send(
-              client.out_buff.c_str(), client.out_buff.length());
-          if (!send_res.has_value())
-            break;
-          const ssize_t bytes = send_res.value();
-          if (bytes == 0)
-            break;
-          client.out_buff.erase(0, static_cast<std::size_t>(bytes));
-        }
-      }
-
-      if (infile.eof()) {
-        client.streaming_file = false;
-        client.out_file_path.clear();
-        if (client.dropping && client.out_buff.empty())
-          disconnect(client_fd);
-        return;
-      }
-    }
-  }
-}
 
 Result<Void> Server::init() {
   // EPoll init
@@ -554,13 +158,23 @@ Result<Void> Server::start(char **envp) {
       if (!resp.keep_alive)
         jt->second.dropping = true;
       jt->second.out_buff = oss.str();
-      client_write(jt->first);
-      if (clients.find(fd) != clients.end() && jt->second.dropping)
-        disconnect(fd);
+      {
+        Result<Void> cw = client_write(jt->first);
+        if (!cw.has_value())
+          return ERR(Void, cw.error());
+      }
+      if (clients.find(fd) != clients.end() && jt->second.dropping) {
+        Result<Void> d = disconnect(fd);
+        if (!d.has_value())
+          return ERR(Void, d.error());
+      }
     }
 
-    for (size_t i = 0; i < clients_to_disconnect.size(); ++i)
-      disconnect(clients_to_disconnect[i]);
+    for (size_t i = 0; i < clients_to_disconnect.size(); ++i) {
+      Result<Void> d = disconnect(clients_to_disconnect[i]);
+      if (!d.has_value())
+        return ERR(Void, d.error());
+    }
 
     if (clients.begin() != clients.end()) {
       const unsigned int session_timeout_ms =
@@ -588,7 +202,11 @@ Result<Void> Server::start(char **envp) {
           if (!resp.keep_alive)
             jt->second.dropping = true;
           jt->second.out_buff = oss.str();
-          client_write(jt->first);
+          {
+            Result<Void> cw = client_write(jt->first);
+            if (!cw.has_value())
+              return ERR(Void, cw.error());
+          }
         }
         cgis_to_reap.insert(cgi);
       } else if (it->second.second->wait_or_reap()) {
@@ -632,9 +250,11 @@ Result<Void> Server::start(char **envp) {
           epoll_timeout = static_cast<long>(cgi_remaining);
       }
     }
-    for (std::set<CgiDelegate *>::const_iterator it = cgis_to_reap.begin();
-         it != cgis_to_reap.end(); ++it)
-      reap_cgi(*it);
+    for (std::set<CgiDelegate *>::const_iterator it = cgis_to_reap.begin(); it != cgis_to_reap.end(); ++it) {
+      Result<Void> rr = reap_cgi(*it);
+      if (!rr.has_value())
+        return ERR(Void, rr.error());
+    }
 
     Result<Events> events_result = epoll.wait(static_cast<int>(epoll_timeout));
     if (!events_result.has_value()) {
@@ -653,16 +273,31 @@ Result<Void> Server::start(char **envp) {
       const FileDescriptor *fd = event->fd;
       std::cerr << utils::debug << "epoll event on fd=" << fd->_fd << std::endl;
       if (listeners.find(fd) != listeners.end()) {
-        new_connection(fd);
+        {
+          Result<Void> nc = new_connection(fd);
+          if (!nc.has_value())
+            return ERR(Void, nc.error());
+        }
       } else if (clients.find(fd) != clients.end()) {
-        if (event->in)
-          client_read(fd, envp);
-        if (event->out)
-          client_write(fd);
+        if (event->in) {
+          Result<Void> cr = client_read(fd, envp);
+          if (!cr.has_value())
+            return ERR(Void, cr.error());
+        }
+        if (event->out) {
+          Result<Void> cw = client_write(fd);
+          if (!cw.has_value())
+            return ERR(Void, cw.error());
+        }
         if (event->err || event->hup || event->rdhup) {
-          if (clients.find(fd) != clients.end())
-            client_read(fd, envp);
-          disconnect(fd);
+          if (clients.find(fd) != clients.end()) {
+            Result<Void> cr = client_read(fd, envp);
+            if (!cr.has_value())
+              return ERR(Void, cr.error());
+          }
+          Result<Void> d = disconnect(fd);
+          if (!d.has_value())
+            return ERR(Void, d.error());
         }
       } else {
         std::map<FileDescriptor const *,
@@ -686,7 +321,9 @@ Result<Void> Server::start(char **envp) {
             std::map<FileDescriptor const *, ClientSession>::iterator jt =
                 clients.find(client_fd);
             if (jt == clients.end()) {
-              reap_cgi(cgi);
+              Result<Void> rr = reap_cgi(cgi);
+              if (!rr.has_value())
+                return ERR(Void, rr.error());
               continue;
             }
             if (output.has_value()) {
@@ -705,8 +342,16 @@ Result<Void> Server::start(char **envp) {
             if (!resp.keep_alive)
               jt->second.dropping = true;
             jt->second.out_buff = oss.str();
-            client_write(client_fd);
-            reap_cgi(cgi);
+            {
+              Result<Void> cw = client_write(client_fd);
+              if (!cw.has_value())
+                return ERR(Void, cw.error());
+            }
+            {
+              Result<Void> rr = reap_cgi(cgi);
+              if (!rr.has_value())
+                return ERR(Void, rr.error());
+            }
           }
         }
       }
@@ -716,16 +361,4 @@ Result<Void> Server::start(char **envp) {
   return OKV;
 }
 
-void Server::reap_cgi(CgiDelegate *cgi) {
-  for (std::map<FileDescriptor const *,
-                std::pair<FileDescriptor const *, CgiDelegate *> >::iterator
-           it = cgis.begin();
-       it != cgis.end();) {
-    if (it->second.second == cgi)
-      cgis.erase(it++);
-    else
-      ++it;
-  }
-  cgi->~CgiDelegate();
-  operator delete(cgi);
-}
+
